LPS-90616 portal-search-engine-adapter-api: Unify duplicate implementation in search and search2 packages
https://issues.liferay.com/browse/LPS-90616
