+++ modules/apps/portal-search/portal-search-engine-adapter-api/bnd.bnd
@@ -7,7 +7,6 @@ Export-Package:\
 	com.liferay.portal.search.engine.adapter.document,\
 	com.liferay.portal.search.engine.adapter.index,\
 	com.liferay.portal.search.engine.adapter.search,\
-	com.liferay.portal.search.engine.adapter.search2,\
 	com.liferay.portal.search.engine.adapter.snapshot
 Import-Package:\
 	com.liferay.portal.kernel.search;version="[7.1.0,8)",\
