+++ modules/apps/portal-search/portal-search-engine-adapter-api/src/main/java/com/liferay/portal/search/engine/adapter/search/SearchSearchResponse.java
@@ -17,6 +17,7 @@ package com.liferay.portal.search.engine.adapter.search;
 import aQute.bnd.annotation.ProviderType;
 
 import com.liferay.portal.kernel.search.Hits;
+import com.liferay.portal.search.hits.SearchHits;
 
 /**
  * @author Michael C. Han
@@ -28,10 +29,28 @@ public class SearchSearchResponse extends BaseSearchResponse {
 		return _hits;
 	}
 
+	public String getScrollId() {
+		return _scrollId;
+	}
+
+	public SearchHits getSearchHits() {
+		return _searchHits;
+	}
+
 	public void setHits(Hits hits) {
 		_hits = hits;
 	}
 
+	public void setScrollId(String scrollId) {
+		_scrollId = scrollId;
+	}
+
+	public void setSearchHits(SearchHits searchHits) {
+		_searchHits = searchHits;
+	}
+
 	private Hits _hits;
+	private String _scrollId;
+	private SearchHits _searchHits;
 
 }
