--- modules/apps/portal-search/portal-search-engine-adapter-api/src/main/java/com/liferay/portal/search/engine/adapter/search2/MultisearchSearchResponse.java
@@ -1,39 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.search.engine.adapter.search2;
-
-import aQute.bnd.annotation.ProviderType;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * @author Michael C. Han
- */
-@ProviderType
-public class MultisearchSearchResponse implements SearchResponse {
-
-	public void addSearchResponse(SearchSearchResponse searchSearchResponse) {
-		_searchSearchResponses.add(searchSearchResponse);
-	}
-
-	public List<SearchSearchResponse> getSearchSearchResponses() {
-		return _searchSearchResponses;
-	}
-
-	private final List<SearchSearchResponse> _searchSearchResponses =
-		new ArrayList<>();
-
-}
