--- modules/apps/portal-search/portal-search-engine-adapter-api/src/main/java/com/liferay/portal/search/engine/adapter/search2/SearchRequestExecutor.java
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.search.engine.adapter.search2;
-
-import aQute.bnd.annotation.ProviderType;
-
-/**
- * @author Michael C. Han
- */
-@ProviderType
-public interface SearchRequestExecutor {
-
-	public CountSearchResponse executeSearchRequest(
-		CountSearchRequest countSearchRequest);
-
-	public MultisearchSearchResponse executeSearchRequest(
-		MultisearchSearchRequest multisearchSearchRequest);
-
-	public SearchSearchResponse executeSearchRequest(
-		SearchSearchRequest searchSearchRequest);
-
-}
