+++ modules/apps/portal-search/portal-search-engine-adapter-api/src/main/resources/com/liferay/portal/search/engine/adapter/packageinfo
@@ -1,1 +1,1 @@
-version 1.3.0
+version 1.2.0
