--- modules/apps/portal-search/portal-search-engine-adapter-api/src/main/resources/com/liferay/portal/search/engine/adapter/search2/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
